[DAL-59] 컨트롤러 분리 작업
