+++ src/main/kotlin/kr/dallyeobom/controller/auth/AuthController.kt
@@ -2,19 +2,15 @@ package kr.dallyeobom.controller.auth
 
 import kr.dallyeobom.controller.auth.request.KakaoLoginRequest
 import kr.dallyeobom.controller.auth.request.KakaoUserCreateRequest
-import kr.dallyeobom.controller.auth.request.NicknameUpdateRequest
 import kr.dallyeobom.controller.auth.response.KakaoLoginResponse
 import kr.dallyeobom.controller.auth.response.NicknameCheckResponse
 import kr.dallyeobom.controller.auth.response.TermsDetailResponse
 import kr.dallyeobom.controller.auth.response.TermsSearchResponse
-import kr.dallyeobom.controller.auth.response.UserInfoResponse
 import kr.dallyeobom.service.UserService
-import kr.dallyeobom.util.LoginUserId
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -39,17 +35,6 @@ class AuthController(
     override fun createKakaoUser(
         @RequestBody @Validated kakaoUserCreateRequest: KakaoUserCreateRequest,
     ) = userService.createUser(kakaoUserCreateRequest)
-
-    @GetMapping("/user")
-    override fun getUserInfo(
-        @LoginUserId userId: Long,
-    ): UserInfoResponse = userService.getUserInfo(userId)
-
-    @PutMapping("/user/nickname")
-    override fun updateNickname(
-        @RequestBody @Validated nicknameUpdateRequest: NicknameUpdateRequest,
-        @LoginUserId userId: Long,
-    ) = userService.updateNickname(nicknameUpdateRequest, userId)
 
     @GetMapping("/terms")
     override fun searchAllTerms(): List<TermsSearchResponse> = userService.searchAllTerms()
