+++ src/main/kotlin/kr/dallyeobom/controller/auth/AuthControllerSpec.kt
@@ -10,14 +10,11 @@ import jakarta.validation.constraints.Size
 import kr.dallyeobom.config.swagger.SwaggerTag
 import kr.dallyeobom.controller.auth.request.KakaoLoginRequest
 import kr.dallyeobom.controller.auth.request.KakaoUserCreateRequest
-import kr.dallyeobom.controller.auth.request.NicknameUpdateRequest
 import kr.dallyeobom.controller.auth.response.KakaoLoginResponse
 import kr.dallyeobom.controller.auth.response.NicknameCheckResponse
 import kr.dallyeobom.controller.auth.response.TermsDetailResponse
 import kr.dallyeobom.controller.auth.response.TermsSearchResponse
-import kr.dallyeobom.controller.auth.response.UserInfoResponse
 import kr.dallyeobom.controller.temporalAuth.response.ServiceTokensResponse
-import kr.dallyeobom.util.LoginUserId
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestBody
@@ -133,50 +130,6 @@ interface AuthControllerSpec {
     fun createKakaoUser(
         @RequestBody @Validated kakaoUserCreateRequest: KakaoUserCreateRequest,
     ): ServiceTokensResponse
-
-    @Operation(
-        summary = "유저 정보를 조회",
-        description = "유저 정보를 조회합니다.",
-        responses = [
-            ApiResponse(
-                responseCode = "200",
-                description = "닉네임과 프로필 정보",
-            ),
-        ],
-    )
-    fun getUserInfo(
-        @LoginUserId userId: Long,
-    ): UserInfoResponse
-
-    @Operation(
-        summary = "닉네임 변경",
-        description = "닉네임을 변경합니다.",
-        responses = [
-            ApiResponse(
-                responseCode = "200",
-                description = "닉네임 변경 성공",
-            ),
-            ApiResponse(
-                responseCode = "409",
-                description = "중복된 닉네임 존재",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        examples = [
-                            ExampleObject(
-                                name = "실패 예시(닉네임 중복)",
-                                value = """{"code": 40900,"errorMessage": "이미 존재하는 닉네임입니다."}""",
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-        ],
-    )
-    fun updateNickname(
-        @RequestBody @Validated nicknameUpdateRequest: NicknameUpdateRequest,
-        @LoginUserId userId: Long,
-    )
 
     @Operation(
         summary = "회원 가입 시 약관 리스트 조회",
